fix: fix unsubscribe

+++ cogito-front/src/app/http-service/thoughts.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
-import { lastValueFrom, Observable, retry, tap } from "rxjs";
+import { lastValueFrom, Observable, retry, Subscription, tap } from "rxjs";
 import { CardType, Tag } from "../types/cards-link";
 import { CogitoStoreService } from "../internal-service/store/cogito-store.service";
 import { map } from "rxjs/operators";
@@ -11,6 +11,7 @@ import { webSocket, WebSocketSubject } from "rxjs/webSocket";
 })
 export class ThoughtsService {
   private ws: WebSocketSubject<{ actionType: "DELETE" | "ADD" | "EDIT"; id: string; card: CardType; }> | undefined;
+  private ws_subscription: Subscription | undefined;
   constructor(
     private httpClient: HttpClient,
     private cogitoStoreService: CogitoStoreService) {
@@ -37,36 +38,42 @@ export class ThoughtsService {
   public connectWebsocket() {
     if (this.ws) {
       console.log("reconnecting websocket");
-
-      this.ws.complete();
-      this.ws.unsubscribe();
+      if (!this.ws.closed) {
+        this.ws.unsubscribe();
+        this.ws_subscription?.unsubscribe();
+      }
       this.ws = this.createWebSocket();
     }
     if (this.ws === undefined) {
       console.log("creating websocket");
       this.ws = this.createWebSocket();
     }
+    if (!this.ws.closed) {
 
-    this.ws.pipe(
-      retry({ count: 99, delay: 1000, resetOnSuccess: true })
-    ).subscribe({
-      next: value => {
-        if (value.actionType === "DELETE") {
-          this.cogitoStoreService.removeCard(value.id);
-        } else if (value.actionType === "ADD") {
-          this.cogitoStoreService.addCard(value.card);
-        } else if (value.actionType === "EDIT") {
-          this.cogitoStoreService.editCard(value.card);
-        }
-      },
-      error: err => {
-        this.cogitoStoreService.websocketStatus(false, err);
-      }, // Called if at any point WebSocket API signals some kind of error.
-      complete: () => {
-        this.cogitoStoreService.websocketStatus(false, null);
-      }// Called when connection is closed (for whatever reason).
-    });
+      this.ws_subscription = this.ws.pipe(
+        retry({ count: 99, delay: 1000, resetOnSuccess: true })
+      ).subscribe({
+        next: value => {
+          this.cogitoStoreService.websocketStatus(true, null);
+          if (value.actionType === "DELETE") {
+            this.cogitoStoreService.removeCard(value.id);
+          } else if (value.actionType === "ADD") {
+            this.cogitoStoreService.addCard(value.card);
+          } else if (value.actionType === "EDIT") {
+            this.cogitoStoreService.editCard(value.card);
+          }
+        },
+        error: err => {
+          this.cogitoStoreService.websocketStatus(false, err);
+        }, // Called if at any point WebSocket API signals some kind of error.
+        complete: () => {
+          this.cogitoStoreService.websocketStatus(false, null);
+        }// Called when connection is closed (for whatever reason).
+      });
+
+    }
   }
+
   public createWebSocket(): WebSocketSubject<{ actionType: "DELETE" | "ADD" | "EDIT"; id: string; card: CardType }> {
     return webSocket<{
       actionType: "DELETE" | "ADD" | "EDIT",
@@ -87,7 +94,7 @@ export class ThoughtsService {
           }
         }
       }
-    )
+    );
 
   }
